use NSUInteger.

+++ App/SVSidebar.m
@@ -53,7 +53,7 @@
 
 - (void)writePageletsHTML:(SVHTMLContext *)context;
 {
-	int savedHeaderLevel = [context currentHeaderLevel];	// probably don't need to save level, but we might as well
+	NSUInteger savedHeaderLevel = [context currentHeaderLevel];	// probably don't need to save level, but we might as well
     [context setCurrentHeaderLevel:4];
     @try
     {
